Project Update. Finish Interactive Menu. New Methods + Documentation. Add Comments.

+++ Sources/TF_Product.cpp
@@ -57,7 +57,7 @@ void TF_Product::discount(double percentage)
 }
 
 /// getter for price
-int TF_Product::getPrice()
+double TF_Product::getPrice()
 {
     return price;
 }
